update README.md and move lib/ to redux/
