+++ examples/with-redux-observable/redux/actionTypes.js
