+++ examples/with-redux-observable/redux/actions.js
