+++ examples/with-redux-observable/redux/epics.js
