+++ examples/with-redux-observable/redux/index.js
