+++ examples/with-redux-observable/redux/reducer.js
